+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to the "Lim Code" extension will be documented in this file.
 
 ## [1.0.87] - 2026-02-12
 
+### Added
+  - 添加了Anthropic格式里自适应思考模式和思考effort参数的配置支持
+
 ### Fixed
   - 修复了不同对话之间，输入框状态不隔离的问题
   - 修复了执行plan后，如果模型回答没有调用任何工具，导致todolist消失的问题
   - 修复了搜索替换工具点击查看差异按钮没有响应的问题
+  - 修复了subagent工具，子代理内部调用ai报错时不重试的问题
+  - 修复了subagent工具，子代理配置的工具列表不被使用的问题
 
 ## [1.0.86] - 2026-02-12